+++ src/nvim/normal.c
@@ -37,6 +37,7 @@
 #include "nvim/keycodes.h"
 #include "nvim/log.h"
 #include "nvim/main.h"
+#include "nvim/mapping.h"
 #include "nvim/mark.h"
 #include "nvim/memline.h"
 #include "nvim/memory.h"
